+++ LuckyoneUI/Core/Defaults.lua
@@ -21,7 +21,6 @@ Private.Defaults = {
 		},
 		misc = {
 			mythicVisibility = false,
-			nameplateClassification = false,
 			dataTextsTweaks = false,
 		},
 		qualityOfLife = {