+++ CHANGELOG.md
@@ -1,4 +1,26 @@
-﻿## Version 4.06 [10-03-2026]
+﻿## Version 4.07 [19-03-2026]
+
+**All WoW Versions:**  
+- ElvUI 15.09+ compatibility  
+- Added a new 4 slot actionbar to the player frame for consumable and trinket tracking  
+- Removed AddOnSkins profile (Outdated)  
+- Removed option for instance-only nameplate classification coloring (its in ElvUI 15.09 and later)  
+- Removed ProjectAzilroka profile (Outdated)  
+- Updated BigWigs profiles  
+- Updated WarpDeplete profile (Font size 11 > 12)  
+
+**Retail:**  
+- Abbreviate enchant names on the character frame by default  
+- Added BuffReminders profile to /lucky config  
+- Added PermoksAccountManager profile to /lucky config  
+- Adjusted PowerBarWidget anchor (Preys etc)  
+- Fixed priest aura indicator overlap (Void Shield / Prayer of Mending)  
+- Updated AyijeCDM profile (Castbar off by default but still configured and ready to use)  
+- Updated BetterCooldownManager profile (Castbar off by default but still configured and ready to use)  
+- Updated Private Auras slightly (Icon size, border scale, anchor position)  
+- Updated WindTools profile (No longer auto skip cutscenes except for myself)  
+
+## Version 4.06 [10-03-2026]
 
 **All WoW Versions:**  
 - Changed back the default ElvUI castbar position to the bottom of the class HUD